+++ packages/stream_feeds/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Next release
-- Update API client
+- Update API client with renaming `addReaction` to `addActivityReaction` and `deleteReaction` to `deleteActivityReaction`.
 
 ## 0.3.0
 - [BREAKING] Renamed `AppLifecycleStateProvider` to `LifecycleStateProvider` and `AppLifecycleState` to `LifecycleState`.